pyproject.toml
[project]
name = "realtime-persiannews-sentiment-llm"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.1",
    "google-genai>=1.53.0",
    "hazm>=0.10.0",
    "jdatetime>=5.2.0",
    "ollama>=0.6.1",
    "pandas>=2.1.0",
    "pika>=1.3.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "scrapy>=2.13.4",
]
